Adding block handling for PowerConverter
I am in need of a fallback if the conversion were to fail. Prior to
this commit, a nil result from the conversion meant that an exception
would be thrown. However, I wanted a fallback option that can be
yielded via the block given to the .convert call.

See the tests for examples.